Simplify code even more
Remove unnecessary type conversions. Now this looks pretty.

+++ mos6502cpu/src/math.rs
@@ -59,15 +59,14 @@ impl Mos6502Cpu {
     #[inline]
     pub(crate) fn execute_sbc_unchecked(&mut self, addressing_mode: &AddressingMode) -> CpuResult {
         let value = self.get_value_from_addressing_mode(addressing_mode)?;
-        let (tmp, first_carry) = (self.registers.a as u16).overflowing_sub(value as u16);
-        let (answer, second_carry) = tmp.overflowing_sub(!self.registers.p.carry as u16);
-        self.update_zero_flag(answer as u8);
-        self.update_negative_flag(answer as u8);
+        let (tmp, first_carry) = self.registers.a.overflowing_sub(value);
+        let (answer, second_carry) = tmp.overflowing_sub(!self.registers.p.carry as u8);
+        self.update_zero_flag(answer);
+        self.update_negative_flag(answer);
         self.registers.p.carry = !(first_carry || second_carry);
         self.registers.p.overflow =
-            ((self.registers.a ^ (answer as u8)) & 0x80) > 0 &&
-                ((self.registers.a ^ value) & 0x80) > 0;
-        self.registers.a = answer as u8;
+            ((self.registers.a ^ answer) & 0x80) > 0 && ((self.registers.a ^ value) & 0x80) > 0;
+        self.registers.a = answer;
         Ok(())
     }
